pyproject.toml
[project]
name = "ip-adapter-palette"
version = "0.1.0"
description = "Ip Adapter for color palette"
authors = [
    { name = "Pierre Colle", email = "piercus@gmail.com" }
]
dependencies = [
    "loguru>=0.7.2",
    "numpy>=1.24.4",
    "Pillow>=10.2.0",
    "refiners[training] @ git+https://github.com/finegrain-ai/refiners.git@d199cd4f242ee33cff8ef9d6776bd171dac39434",
    "pydantic>=2.6.1",
    "scikit-learn>=1.3.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pyright==1.1.349",
    "ruff>=0.2.0",
    "pytest>=8.0.0",
    "transformers>=4.38.0",
    "diffusers>=0.26.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/palette_adapter"]
